Added `GuildCommand.kt`
Which includes:
- GuildAddStickerExecutor (to add stickers)
- GuildRemoveStickerExecutor (to remove stickers)
- GuildAddEmojiExecutor (to replace `+addemoji`)
- GuildRemoveEmojiExecutor (to replace `+removeemoji`)

The "remove" versions of the stickers and emojis has an independent class to manage the "remove" message.

Also added `MessageStickerCommand.kt` to ~~steal~~ get other users stickers.

Also² updated the `OptionReference.kt` adding a filter to search images history or not.

Also³ fixed the line 229 at `UnleashedCommandManager.kt` as @MrPowerGamerBR said.